tools: Move algos to algorithms.py
